fix(activation): cast the vaddq literal to its lane type to avoid the MVE _Generic undef trap (#306)
The MVE tanh gather paths passed a bare `1U` literal to the polymorphic `vaddq`. On arm-none-eabi `1U` is `unsigned int` while `uint32_t` is `long unsigned int`, and GCC 13's `__ARM_mve_coerce_i_scalar` accepts neither `unsigned int` nor any of its typedefs - so the call fell through to the `__ARM_undef` link trap instead of failing at compile time. Casting the literal to each path's lane type resolves the dispatch.

Affects GCC 13 only: 14 and 15 dropped `vaddq` from their `_Generic` macro set, and their objects are byte-identical before and after this change. The published v7.29.2 and v7.28.0 cortex-m55 GCC archives each carry four objects with the unresolvable symbol (activation and lstm_step, f16 and f32); the ATfE archives for those tags are clean. Verified across GCC 13.2 through 15.2 by real whole-archive links, and by QEMU execution over the LUT boundary, NaN and Inf giving bit-identical output on every configuration.

The release smoke gate could not have caught this - it links with `--gc-sections` and `--unresolved-symbols=ignore-all`. PR #307 makes that link strict and adds a per-major GCC matrix so the class is gated going forward.

Closes #305

+++ Include/Internal/arm_nn_activation_flt.h
@@ -233,7 +233,8 @@ __STATIC_INLINE float32x4_t arm_nn_vtanh_lut_direct_mve_f32(float32x4_t x)
     idx = vminq(idx, vdupq_n_u32((uint32_t)ARM_NN_TANH_F32_LUT_MAX_IDX));
     const float32x4_t frac = vsubq(t, vcvtq_f32_u32(idx));
     const float32x4_t y0 = vldrwq_gather_shifted_offset((const float32_t *)arm_nn_tanh_lut384_f32, idx);
-    const float32x4_t y1 = vldrwq_gather_shifted_offset((const float32_t *)arm_nn_tanh_lut384_f32, vaddq(idx, 1U));
+    const float32x4_t y1 =
+        vldrwq_gather_shifted_offset((const float32_t *)arm_nn_tanh_lut384_f32, vaddq(idx, (uint32_t)1U));
     float32x4_t y = vfmaq(y0, vsubq(y1, y0), frac);
     y = vpselq(vdupq_n_f32(1.0f), y, sat_p);
     return vnegq_m(y, y, vcmpltq(x, 0.0f));
@@ -383,7 +384,8 @@ __STATIC_INLINE float16x8_t arm_nn_vtanh_lut_direct_mve_f16(float16x8_t x)
     idx = vminq(idx, vdupq_n_u16(255U));
     const float16x8_t frac = vsubq(t, vcvtq_f16_u16(idx));
     const float16x8_t y0 = vldrhq_gather_shifted_offset((const float16_t *)arm_nn_tanh_lut256_f16, idx);
-    const float16x8_t y1 = vldrhq_gather_shifted_offset((const float16_t *)arm_nn_tanh_lut256_f16, vaddq(idx, 1U));
+    const float16x8_t y1 =
+        vldrhq_gather_shifted_offset((const float16_t *)arm_nn_tanh_lut256_f16, vaddq(idx, (uint16_t)1U));
     float16x8_t y = vfmaq(y0, vsubq(y1, y0), frac);
     y = vpselq(vdupq_n_f16((float16_t)1.0f), y, sat_p);
     return vnegq_m(y, y, vcmpltq(x, (float16_t)0.0f));
